feat - 장바구니 담기, 리스트 출력

+++ src/main/java/animal_shop/shop/cart/dto/CartDetailDTOResponse.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Builder
 public class CartDetailDTOResponse {
     private List<CartDetailDTO> cartDetailDTOList;
-    private int total_count;
+    private Long total_count;
 
 }
